Split up PartitionDefinition
