+++ src/Components/PartitionDefinition.php
@@ -5,14 +5,8 @@ declare(strict_types=1);
 namespace PhpMyAdmin\SqlParser\Components;
 
 use PhpMyAdmin\SqlParser\Component;
-use PhpMyAdmin\SqlParser\Parseable;
-use PhpMyAdmin\SqlParser\Parser;
-use PhpMyAdmin\SqlParser\Parsers\ArrayObjs;
-use PhpMyAdmin\SqlParser\Parsers\OptionsArrays;
-use PhpMyAdmin\SqlParser\TokensList;
-use PhpMyAdmin\SqlParser\TokenType;
+use PhpMyAdmin\SqlParser\Parsers\PartitionDefinitions;
 
-use function implode;
 use function trim;
 
 /**
@@ -20,7 +14,7 @@ use function trim;
  *
  * Used for parsing `CREATE TABLE` statement.
  */
-final class PartitionDefinition implements Component, Parseable
+final class PartitionDefinition implements Component
 {
     /**
      * All field options.
@@ -109,125 +103,13 @@ final class PartitionDefinition implements Component, Parseable
      */
     public $options;
 
-    /**
-     * @param Parser               $parser  the parser that serves as context
-     * @param TokensList           $list    the list of tokens that are being parsed
-     * @param array<string, mixed> $options parameters for parsing
-     */
-    public static function parse(Parser $parser, TokensList $list, array $options = []): PartitionDefinition
-    {
-        $ret = new static();
-
-        /**
-         * The state of the parser.
-         *
-         * Below are the states of the parser.
-         *
-         *      0 -------------[ PARTITION | SUBPARTITION ]------------> 1
-         *
-         *      1 -----------------------[ name ]----------------------> 2
-         *
-         *      2 ----------------------[ VALUES ]---------------------> 3
-         *
-         *      3 ---------------------[ LESS THAN ]-------------------> 4
-         *      3 ------------------------[ IN ]-----------------------> 4
-         *
-         *      4 -----------------------[ expr ]----------------------> 5
-         *
-         *      5 ----------------------[ options ]--------------------> 6
-         *
-         *      6 ------------------[ subpartitions ]------------------> (END)
-         */
-        $state = 0;
-
-        for (; $list->idx < $list->count; ++$list->idx) {
-            /**
-             * Token parsed at this moment.
-             */
-            $token = $list->tokens[$list->idx];
-
-            // End of statement.
-            if ($token->type === TokenType::Delimiter) {
-                break;
-            }
-
-            // Skipping whitespaces and comments.
-            if (($token->type === TokenType::Whitespace) || ($token->type === TokenType::Comment)) {
-                continue;
-            }
-
-            if ($state === 0) {
-                $ret->isSubpartition = ($token->type === TokenType::Keyword) && ($token->keyword === 'SUBPARTITION');
-                $state = 1;
-            } elseif ($state === 1) {
-                $ret->name = $token->value;
-
-                // Looking ahead for a 'VALUES' keyword.
-                // Loop until the end of the partition name (delimited by a whitespace)
-                while ($nextToken = $list->tokens[++$list->idx]) {
-                    if ($nextToken->type !== TokenType::None) {
-                        break;
-                    }
-
-                    $ret->name .= $nextToken->value;
-                }
-
-                $idx = $list->idx--;
-                // Get the first token after the white space.
-                $nextToken = $list->tokens[++$idx];
-
-                $state = ($nextToken->type === TokenType::Keyword)
-                    && ($nextToken->value === 'VALUES')
-                    ? 2 : 5;
-            } elseif ($state === 2) {
-                $state = 3;
-            } elseif ($state === 3) {
-                $ret->type = $token->value;
-                $state = 4;
-            } elseif ($state === 4) {
-                if ($token->value === 'MAXVALUE') {
-                    $ret->expr = $token->value;
-                } else {
-                    $ret->expr = Expression::parse(
-                        $parser,
-                        $list,
-                        [
-                            'parenthesesDelimited' => true,
-                            'breakOnAlias' => true,
-                        ],
-                    );
-                }
-
-                $state = 5;
-            } elseif ($state === 5) {
-                $ret->options = OptionsArrays::parse($parser, $list, static::$partitionOptions);
-                $state = 6;
-            } elseif ($state === 6) {
-                if (($token->type === TokenType::Operator) && ($token->value === '(')) {
-                    $ret->subpartitions = ArrayObjs::parse(
-                        $parser,
-                        $list,
-                        ['type' => self::class],
-                    );
-                    ++$list->idx;
-                }
-
-                break;
-            }
-        }
-
-        --$list->idx;
-
-        return $ret;
-    }
-
     public function build(): string
     {
         if ($this->isSubpartition) {
             return trim('SUBPARTITION ' . $this->name . ' ' . $this->options);
         }
 
-        $subpartitions = empty($this->subpartitions) ? '' : ' ' . self::buildAll($this->subpartitions);
+        $subpartitions = empty($this->subpartitions) ? '' : ' ' . PartitionDefinitions::buildAll($this->subpartitions);
 
         return trim(
             'PARTITION ' . $this->name
@@ -235,12 +117,6 @@ final class PartitionDefinition implements Component, Parseable
             . (! empty($this->options) && ! empty($this->type) ? '' : ' ')
             . $this->options . $subpartitions,
         );
-    }
-
-    /** @param PartitionDefinition[] $component the component to be built */
-    public static function buildAll(array $component): string
-    {
-        return "(\n" . implode(",\n", $component) . "\n)";
     }
 
     public function __toString(): string
