+++ tests/Components/PartitionDefinitionTest.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\PartitionDefinition;
 use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Parsers\PartitionDefinitions;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class PartitionDefinitionTest extends TestCase
 {
     public function testParse(): void
     {
-        $component = PartitionDefinition::parse(
+        $component = PartitionDefinitions::parse(
             new Parser(),
             $this->getTokensList('PARTITION p0 VALUES LESS THAN(1990)'),
         );
@@ -24,7 +24,7 @@ class PartitionDefinitionTest extends TestCase
 
     public function testParseNameWithUnderscore(): void
     {
-        $component = PartitionDefinition::parse(
+        $component = PartitionDefinitions::parse(
             new Parser(),
             $this->getTokensList('PARTITION 2017_12 VALUES LESS THAN (\'2018-01-01 00:00:00\') ENGINE = MyISAM'),
         );
